chore: :recycle: rename files to kebab-case
