+++ src/components/emoji-picker-button.tsx
